Major file restructuring and renaming overhaul
